SUMMARY.md can split book into parts
Parts should be split with a level two heading, or a paragraph. Part
heading can be a link. Chapter numbering is reset for each part.

The summary object adds a new property “parts” which exposes an array
of objects with title, link, level and chapters. The level property is
always an empty string.

To prevent having to rewrite any templates, the summary object also
exposes a chapters property which concatenates all of the parts as
simple chapters. The parts get a level number, and child chapters
levels are also adjusted