Number Single Level books as multiple levels.
Fix #1301

This makes it so that a single level book will have chapters show as
"1.", "2." and so on, making single part books appear the same as
before.

Note that the smallest way to make this change causes the `level`
property to be an integer for single part books. I'm not 100% sure this
is desirable, but so far it didn't seem to cause complications.