Fix scroll-padding issue
In summary: every time we'd type in the input, the browser would try to scroll to the input. But because the input is shown _above_ the scrollview's `scroll-padding-top`, it jumped up a tiny bit with every character typed. Instead of attaching a `scroll-padding` to the viewport, we now have a `scroll-margin` on elements we're actually scrolling to. This was even in place already, so took no extra effort.