Fix Search results are not clickable
The root cause for this was that we instantiated two search popovers on sites without a header, since these sites still have a mobile header with its own search popover in addition to the sidebar search popover.

Add a `visibility` check to only render the popover when appropriate.

Also cleans up the mobile sidebar to not repeat the logo & search button unncessarily.