Fix `@gitbook/icons` icon rendering
The icon rendering has been broken since v0.3.2 of this package. The `src/style.css` was no longer included on following versions. That CSS file set `background-color: currentColor` on all `.gb-icon` elements. With it missing, none of the icons rendered out of the box. It has gone unnoticed until now because GBX uses `0.3.1` and has never bumped to the latest versions.

This patch fixes the rendering by including a `backgroundColor: currentColor` assignment on the icon svg itself, rendering the `style.css` unnecessary, so we can safely remove it. Icons can still override the `backgroundColor` as they have been doing in GBX.